% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_risk_reward}
\alias{plot_risk_reward}
\title{Display a project risk-reward bubble plot}
\usage{
plot_risk_reward(data)
}
\arguments{
\item{data}{A dataset with columns Success_prob, Reward, Effort, Group, and Project (name)}
}
\value{
ggplot2 object
}
\description{
Display a project risk-reward bubble plot
}
\examples{
library(tibble)
data <- tibble::tibble(Success_prob=c(0.1,0.9),Reward=c(0.3,0.7),
       Effort=c(0.3,0.5),Project=c("A","B"),Group=c("Old","New"))
plot_risk_reward(data)
}
